pyproject.toml
[tool.poetry]
name = "meme-generator"
version = "0.1.0"
description = ""
authors = ["Mark Clare <mark.clare@kindredgroup.com>"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.3"
requests = "^2.28.2"
python-docx = "^0.8.11"
pandas = "^1.5.3"
Pillow = "^9.4.0"


[tool.poetry.dev-dependencies]
bandit = "*"
pre-commit = "*"
gitdb = "*"
isort = "*"
black = "*"
flake8 = "*"
pylint = "*"
mypy = "*"
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
